requirements.txt
tensorflow
opencv-python
numpy
pandas
scikit-learn
mtcnn
deepface
fer
streamlit
Pillow